allow for larger min write than 255.

+++ boot/bootutil/src/bootutil_public.c
@@ -357,7 +357,7 @@ boot_write_trailer(const struct flash_area *fap, uint32_t off,
         const uint8_t *inbuf, uint8_t inlen)
 {
     uint8_t buf[BOOT_MAX_ALIGN];
-    uint8_t align;
+    uint32_t align;
     uint8_t erased_val;
     int rc;
